update neovim config

+++ .local/share/nvim/plugged/nvim-tree.lua/lua/nvim-tree/fs.lua
@@ -1,6 +1,5 @@
 local api = vim.api
 local luv = vim.loop
-local open_mode = luv.constants.O_CREAT + luv.constants.O_WRONLY + luv.constants.O_TRUNC
 
 local utils = require'nvim-tree.utils'
 local view = require'nvim-tree.view'
@@ -29,13 +28,10 @@ local function create_file(file)
       return
     end
   end
-  luv.fs_open(file, "w", open_mode, vim.schedule_wrap(function(err, fd)
+  luv.fs_open(file, "w", 420, vim.schedule_wrap(function(err, fd)
     if err then
       api.nvim_err_writeln('Couldn\'t create file '..file)
     else
-      -- FIXME: i don't know why but libuv keeps creating file with executable permissions
-      -- this is why we need to chmod to default file permissions
-      luv.fs_chmod(file, 420)
       luv.fs_close(fd)
       events._dispatch_file_created(file)
       lib.refresh_tree()
@@ -58,60 +54,81 @@ function M.create(node)
     node = {
       absolute_path = lib.Tree.cwd,
       entries = lib.Tree.entries,
+      open = true,
     }
   end
 
+  local node_is_open = vim.g.nvim_tree_create_in_closed_folder == 1 or node.open
+
   local add_into
-  if node.entries ~= nil then
+  if node.entries ~= nil and node_is_open then
     add_into = utils.path_add_trailing(node.absolute_path)
   else
-    add_into = node.absolute_path:sub(0, -(#node.name + 1))
+    add_into = node.absolute_path:sub(0, -(#(node.name or '') + 1))
   end
 
-  local ans = vim.fn.input('Create file '..add_into)
+  local ans = vim.fn.input('Create file ', add_into)
   utils.clear_prompt()
-  if not ans or #ans == 0 then return end
+  if not ans or #ans == 0 or luv.fs_access(ans, 'r') then return end
 
-  if not ans:match(utils.path_separator) then
-    return create_file(add_into..ans)
-  end
-
-  -- create a foler for each element until / and create a file when element is not ending with /
-  -- if element is ending with / and it's the last element, we need to manually refresh
-  local relpath = ''
+  -- create a folder for each path element if the folder does not exist
+  -- if the answer ends with a /, create a file for the last entry
+  local is_last_path_file = not ans:match(utils.path_separator..'$')
+  local path_to_create = ''
   local idx = 0
 
-  local num_entries = get_num_entries(utils.path_split(ans))
-  local first_entry
+  local num_entries = get_num_entries(utils.path_split(utils.path_remove_trailing(ans)))
   for path in utils.path_split(ans) do
-    if idx == 0 then
-      first_entry = add_into..relpath..path
-    end
     idx = idx + 1
-    relpath = relpath..path
-    local abs_path = add_into..relpath
-    if relpath:match('.*'..utils.path_separator..'$') then
-      local success = luv.fs_mkdir(abs_path, 493)
-      if not success then
-        api.nvim_err_writeln('Could not create folder '..abs_path)
-        return
-      end
-      if idx == num_entries then
-        events._dispatch_folder_created(abs_path)
-        api.nvim_out_write('Folder '..abs_path..' was properly created\n')
-        lib.refresh_tree()
-      end
+    local p = utils.path_remove_trailing(path)
+    if #path_to_create == 0 and vim.fn.has('win32') == 1 then
+      path_to_create = utils.path_join({p, path_to_create})
     else
-      create_file(abs_path)
+      path_to_create = utils.path_join({path_to_create, p})
+    end
+    if is_last_path_file and idx == num_entries then
+      create_file(path_to_create)
+    elseif not luv.fs_access(path_to_create, "r") then
+      local success = luv.fs_mkdir(path_to_create, 493)
+      if not success then
+        api.nvim_err_writeln('Could not create folder '..path_to_create)
+        break
+      end
     end
   end
-  focus_file(first_entry:sub(0, #first_entry - 1))
+  api.nvim_out_write(ans..' was properly created\n')
+  events._dispatch_folder_created(ans)
+  lib.refresh_tree()
+  focus_file(ans)
 end
 
 local function clear_buffer(absolute_path)
+  local bufs = vim.fn.getbufinfo({bufloaded = 1, buflisted = 1})
+  for _, buf in pairs(bufs) do
+    if buf.name == absolute_path then
+      if buf.hidden == 0 and #bufs > 1 then
+        local winnr = api.nvim_get_current_win()
+        api.nvim_set_current_win(buf.windows[1])
+        vim.cmd(':bn')
+        api.nvim_set_current_win(winnr)
+      end
+      vim.api.nvim_buf_delete(buf.bufnr, {})
+      if buf.windows[1] then
+        vim.api.nvim_win_close(buf.windows[1], true)
+      end
+      return
+    end
+  end
+end
+
+local function rename_loaded_buffers(old_name, new_name)
   for _, buf in pairs(api.nvim_list_bufs()) do
-    if vim.fn.bufloaded(buf) == 1 and api.nvim_buf_get_name(buf) == absolute_path then
-      api.nvim_command(':bd! '..buf)
+    if api.nvim_buf_is_loaded(buf) then
+      if api.nvim_buf_get_name(buf) == old_name then
+        api.nvim_buf_set_name(buf, new_name)
+        -- to avoid the 'overwrite existing file' error message on write
+        vim.api.nvim_buf_call(buf, function() vim.cmd("silent! w!") end)
+      end
     end
   end
 end
@@ -165,6 +182,15 @@ local function do_copy(source, destination)
     if not success then return success, msg end
   end
 
+  return true
+end
+
+local function do_cut(source, destination)
+  local success = luv.fs_rename(source, destination)
+  if not success then
+    return success
+  end
+  rename_loaded_buffers(source, destination)
   return true
 end
 
@@ -239,7 +265,7 @@ function M.remove(node)
   local ans = utils.get_user_input_char()
   utils.clear_prompt()
   if ans:match('^y') then
-    if node.entries ~= nil then
+    if node.entries ~= nil and not node.link_to then
       local success = remove_dir(node.absolute_path)
       if not success then
         return api.nvim_err_writeln('Could not remove '..node.name)
@@ -266,22 +292,20 @@ function M.rename(with_sub)
     local abs_path = with_sub and node.absolute_path:sub(0, namelen * (-1) -1) or node.absolute_path
     local new_name = vim.fn.input("Rename " ..node.name.. " to ", abs_path)
     utils.clear_prompt()
-    if not new_name or #new_name == 0 then return end
+    if not new_name or #new_name == 0 then
+      return
+    end
+    if luv.fs_access(new_name, 'R') then
+      utils.warn("Cannot rename: file already exists")
+      return
+    end
 
     local success = luv.fs_rename(node.absolute_path, new_name)
     if not success then
       return api.nvim_err_writeln('Could not rename '..node.absolute_path..' to '..new_name)
     end
     api.nvim_out_write(node.absolute_path..' ➜ '..new_name..'\n')
-    for _, buf in pairs(api.nvim_list_bufs()) do
-      if api.nvim_buf_is_loaded(buf) then
-        if api.nvim_buf_get_name(buf) == node.absolute_path then
-          api.nvim_buf_set_name(buf, new_name)
-          -- to avoid the 'overwrite existing file' error message on write
-          vim.api.nvim_buf_call(buf, function() vim.cmd("silent! w!") end)
-        end
-      end
-    end
+    rename_loaded_buffers(node.absolute_path, new_name)
     events._dispatch_node_renamed(abs_path, new_name)
     lib.refresh_tree()
   end
@@ -297,7 +321,7 @@ end
 
 function M.paste(node)
   if clipboard.move[1] ~= nil then
-    return do_paste(node, 'move', luv.fs_rename)
+    return do_paste(node, 'move', do_cut)
   end
 
   return do_paste(node, 'copy', do_copy)
